Report the input string when throwing FileLoadException
